+++ delicious-climbing/src/main/java/sangmyungdae/deliciousclimbing/domain/entity/TbPost.java
@@ -5,7 +5,6 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import sangmyungdae.deliciousclimbing.domain.enums.BoardType;
-import sangmyungdae.deliciousclimbing.dto.Post;
 
 import javax.persistence.*;
 import java.util.ArrayList;
