+++ delicious-climbing/src/main/java/sangmyungdae/deliciousclimbing/service/CommunityService.java
@@ -10,7 +10,12 @@ import sangmyungdae.deliciousclimbing.domain.entity.TbFile;
 import sangmyungdae.deliciousclimbing.domain.entity.TbPost;
 import sangmyungdae.deliciousclimbing.domain.entity.TbUser;
 import sangmyungdae.deliciousclimbing.domain.enums.BoardType;
-import sangmyungdae.deliciousclimbing.dto.*;
+import sangmyungdae.deliciousclimbing.dto.community.Comment;
+import sangmyungdae.deliciousclimbing.dto.community.CommentDto;
+import sangmyungdae.deliciousclimbing.dto.community.File;
+import sangmyungdae.deliciousclimbing.dto.community.FileDto;
+import sangmyungdae.deliciousclimbing.dto.mate.Post;
+import sangmyungdae.deliciousclimbing.dto.mate.PostDto;
 import sangmyungdae.deliciousclimbing.repository.CommentRepository;
 import sangmyungdae.deliciousclimbing.repository.FileRepository;
 import sangmyungdae.deliciousclimbing.repository.PostRepository;
