+++ delicious-climbing/src/test/java/sangmyungdae/deliciousclimbing/service/CommunityServiceTest.java
@@ -8,15 +8,14 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.transaction.annotation.Transactional;
 import sangmyungdae.deliciousclimbing.domain.entity.TbUser;
 import sangmyungdae.deliciousclimbing.domain.enums.BoardType;
 import sangmyungdae.deliciousclimbing.domain.enums.Gender;
 import sangmyungdae.deliciousclimbing.domain.enums.LoginType;
-import sangmyungdae.deliciousclimbing.dto.Comment;
-import sangmyungdae.deliciousclimbing.dto.CommentDto;
-import sangmyungdae.deliciousclimbing.dto.Post;
-import sangmyungdae.deliciousclimbing.dto.PostDto;
+import sangmyungdae.deliciousclimbing.dto.community.Comment;
+import sangmyungdae.deliciousclimbing.dto.community.CommentDto;
+import sangmyungdae.deliciousclimbing.dto.mate.Post;
+import sangmyungdae.deliciousclimbing.dto.mate.PostDto;
 import sangmyungdae.deliciousclimbing.repository.CommentRepository;
 import sangmyungdae.deliciousclimbing.repository.FileRepository;
 import sangmyungdae.deliciousclimbing.repository.PostRepository;
